Attribute Macro jf_utils_derive::tagged_blob

#[tagged_blob]

Derive serdes for a type which serializes as a binary blob.

This macro can be used to easily derive friendly serde implementations for a binary type which implements CanonicalSerialize and CanonicalDeserialize. This is useful for cryptographic primitives and other types which do not have a human-readable serialization, but which may be embedded in structs with a human-readable serialization. The serde implementations derived by this macro will serialize the type as bytes for binary encodings and as base 64 for human readable encodings.

Specifically, this macro does 4 things when applied to a type definition:

It adds #[derive(Serialize, Deserialize)] to the type definition, along with serde attributes to serialize using the TaggedBlob serialization helper.

It creates an implementation of Tagged for the type using the specified tag. This tag will be used to identify base 64 strings which represent this type in human-readable encodings.

It creates an implementation of TryFrom<TaggedBlob<T>> for the type T, which is needed to make the serde(try_from) attribute work.

It creates implementations of Display and FromStr using tagged base 64 as a display format. This allows tagged blob types to be conveniently displayed and read to and from user interfaces in a manner consistent with how they are serialized.

Usage example:

#[macro_use] extern crate jf_utils_derive;
use ark_serialize::*;

#[tagged_blob("NUL")]
#[derive(Clone, CanonicalSerialize, CanonicalDeserialize, /* any other derives */)]
pub struct Nullifier(
    // This type can only be serialied as an opaque, binary blob using ark_serialize.
    pub(crate) ark_bls12_381::Fr,
);

The type Nullifier can now be serialized as binary:

bincode::serialize(&nullifier).unwrap();

or as base64:

serde_json::to_string(&nullifier).unwrap();

which will produce a tagged base64 string like “NUL~8oaujwbov8h4eEq7HFpqW6mIXhVbtJGxLUgiKrGpMCoJ”.